fold internal/spinner into internal/display
Move spinner.go, model.go, and format.go from internal/spinner/ into
internal/display/spinner.go. The spinner is display-layer infrastructure
and belongs alongside the rest of the display package.

Changes:
- internal/display/spinner.go: combined from all three spinner files
  - SpinnerShouldShow (was ShouldShow)
  - SpinnerRun (was Run)
  - CompletionInfo type (was in format.go)
  - spinnerModel / newSpinnerModel (was model / newModel)
  - FormatCompletionText inlined as c.ProviderID in View()
  - internal helpers renamed to avoid collisions (spinnerCheckStyle,
    spinnerErrStyle, removeStringFromSlice)
- internal/display/spinner_test.go: combined from three test files,
  updated for new names
- internal/cli/root.go: spinner.* → display.Spinner*
- internal/cli/route.go: spinner.* → display.Spinner*
- internal/cli/root_test.go: spinner.CompletionInfo → display.CompletionInfo
- internal/spinner/: deleted
