+++ internal/display/spinner_test.go
@@ -1,4 +1,4 @@
-package spinner
+package display
 
 import (
 	"strings"
@@ -7,9 +7,49 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 )
 
-func TestNewModel(t *testing.T) {
+func TestSpinnerShouldShow(t *testing.T) {
+	tests := []struct {
+		name   string
+		quiet  bool
+		json   bool
+		nonTTY bool
+		want   bool
+	}{
+		{"interactive", false, false, false, true},
+		{"quiet mode", true, false, false, false},
+		{"json mode", false, true, false, false},
+		{"both quiet and json", true, true, false, false},
+		{"non-TTY (piped)", false, false, true, false},
+		{"quiet and non-TTY", true, false, true, false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := SpinnerShouldShow(tt.quiet, tt.json, tt.nonTTY)
+			if got != tt.want {
+				t.Errorf("SpinnerShouldShow(quiet=%v, json=%v, nonTTY=%v) = %v, want %v",
+					tt.quiet, tt.json, tt.nonTTY, got, tt.want)
+			}
+		})
+	}
+}
+
+func TestSpinnerRunEmptyProviders(t *testing.T) {
+	called := false
+	err := SpinnerRun([]string{}, func(onComplete func(CompletionInfo)) {
+		called = true
+	})
+	if err != nil {
+		t.Errorf("SpinnerRun with empty providers returned error: %v", err)
+	}
+	if !called {
+		t.Error("expected fetchFn to be called even with empty providers")
+	}
+}
+
+func TestNewSpinnerModel(t *testing.T) {
 	providers := []string{"claude", "copilot", "gemini"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
 	if len(m.inflight) != 3 {
 		t.Errorf("expected 3 inflight, got %d", len(m.inflight))
@@ -22,16 +62,15 @@ func TestNewModel(t *testing.T) {
 	}
 }
 
-func TestModelUpdateCompletion(t *testing.T) {
+func TestSpinnerModelUpdateCompletion(t *testing.T) {
 	providers := []string{"claude", "copilot", "gemini"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
-	// Complete one provider
-	updated, cmd := m.Update(completionMsg{
+	updated, cmd := m.Update(spinnerCompletionMsg{
 		ProviderID: "copilot",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	if len(m.inflight) != 2 {
 		t.Errorf("expected 2 inflight, got %d", len(m.inflight))
@@ -50,23 +89,21 @@ func TestModelUpdateCompletion(t *testing.T) {
 	}
 }
 
-func TestModelUpdateAllComplete(t *testing.T) {
+func TestSpinnerModelUpdateAllComplete(t *testing.T) {
 	providers := []string{"claude", "copilot"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
-	// Complete first
-	updated, _ := m.Update(completionMsg{
+	updated, _ := m.Update(spinnerCompletionMsg{
 		ProviderID: "claude",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
-	// Complete second — should trigger quit
-	updated, cmd := m.Update(completionMsg{
+	updated, cmd := m.Update(spinnerCompletionMsg{
 		ProviderID: "copilot",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	if len(m.inflight) != 0 {
 		t.Errorf("expected 0 inflight, got %d", len(m.inflight))
@@ -79,35 +116,33 @@ func TestModelUpdateAllComplete(t *testing.T) {
 	}
 }
 
-func TestModelUpdateDuplicateCompletion(t *testing.T) {
+func TestSpinnerModelUpdateDuplicateCompletion(t *testing.T) {
 	providers := []string{"claude", "copilot"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
-	// Complete claude
-	updated, _ := m.Update(completionMsg{
+	updated, _ := m.Update(spinnerCompletionMsg{
 		ProviderID: "claude",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
-	// Complete claude again (duplicate) — should be ignored
-	updated, _ = m.Update(completionMsg{
+	updated, _ = m.Update(spinnerCompletionMsg{
 		ProviderID: "claude",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	if len(m.completions) != 1 {
 		t.Errorf("expected 1 completion (duplicate ignored), got %d", len(m.completions))
 	}
 }
 
-func TestModelUpdateCtrlC(t *testing.T) {
+func TestSpinnerModelUpdateCtrlC(t *testing.T) {
 	providers := []string{"claude"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
 	updated, cmd := m.Update(tea.KeyMsg{Type: tea.KeyCtrlC})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	if !m.quitting {
 		t.Error("expected quitting=true on ctrl+c")
@@ -117,13 +152,12 @@ func TestModelUpdateCtrlC(t *testing.T) {
 	}
 }
 
-func TestModelViewInflight(t *testing.T) {
+func TestSpinnerModelViewInflight(t *testing.T) {
 	providers := []string{"claude", "copilot", "gemini"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
 	view := m.View()
 
-	// Each provider should have its own line with a spinner
 	for _, p := range providers {
 		if !strings.Contains(view, p) {
 			t.Errorf("expected provider %q in view, got %q", p, view)
@@ -135,70 +169,64 @@ func TestModelViewInflight(t *testing.T) {
 	}
 }
 
-func TestModelViewPartialCompletion(t *testing.T) {
+func TestSpinnerModelViewPartialCompletion(t *testing.T) {
 	providers := []string{"claude", "copilot", "gemini"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
-	updated, _ := m.Update(completionMsg{
+	updated, _ := m.Update(spinnerCompletionMsg{
 		ProviderID: "copilot",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	view := m.View()
 
-	// All providers should still be visible, each on its own line
 	for _, p := range providers {
 		if !strings.Contains(view, p) {
 			t.Errorf("expected provider %q in view, got %q", p, view)
 		}
 	}
-	// Copilot's line should have a checkmark
 	if !strings.Contains(view, "✓") {
 		t.Errorf("expected checkmark for completed copilot, got %q", view)
 	}
 }
 
-func TestModelViewAllDone(t *testing.T) {
+func TestSpinnerModelViewAllDone(t *testing.T) {
 	providers := []string{"claude"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
-	updated, _ := m.Update(completionMsg{
+	updated, _ := m.Update(spinnerCompletionMsg{
 		ProviderID: "claude",
 		Success:    true,
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	view := m.View()
 
-	// Final view should be empty — spinner is transient progress UI
 	if view != "" {
 		t.Errorf("expected empty final view, got %q", view)
 	}
 }
 
-func TestModelViewShowsFailures(t *testing.T) {
+func TestSpinnerModelViewShowsFailures(t *testing.T) {
 	providers := []string{"claude", "cursor"}
-	m := newModel(providers)
+	m := newSpinnerModel(providers)
 
-	// Complete cursor with failure
-	updated, _ := m.Update(completionMsg{
+	updated, _ := m.Update(spinnerCompletionMsg{
 		ProviderID: "cursor",
 		Success:    false,
 		Error:      "not configured",
 	})
-	m = updated.(model)
+	m = updated.(spinnerModel)
 
 	view := m.View()
 
-	// Failed provider should show with ✗
 	if !strings.Contains(view, "✗") {
 		t.Errorf("expected failure indicator ✗ in view, got %q", view)
 	}
 	if !strings.Contains(view, "cursor") {
 		t.Errorf("expected failed provider cursor in view, got %q", view)
 	}
-	// Should still show claude as inflight
 	if !strings.Contains(view, "claude") {
 		t.Errorf("expected inflight provider claude in view, got %q", view)
 	}
