--- internal/spinner/format.go
@@ -1,13 +1,0 @@
-package spinner
-
-// CompletionInfo describes a completed provider fetch.
-type CompletionInfo struct {
-	ProviderID string
-	Success    bool
-	Error      string
-}
-
-// FormatCompletionText formats the text portion of a completion line (without symbol).
-func FormatCompletionText(info CompletionInfo) string {
-	return info.ProviderID
-}
